==> notebooks/01_cleaning_and_filtering.ipynb <==
## 1. Loading Dataset and Overview

# Import necessary libraries
import pandas as pd

### 1.1 Loading Dataset

# Load the dataset
raw_data_path = '../data/product_raw.csv'
raw_data = pd.read_csv(raw_data_path)

# Display the dataset
raw_data


### 1.2 Data overview

# Display the dataset shape
print("Shape of DataFrame:", raw_data.shape)


# Display the number of unique values in each column
print("Number of unique values in each column:\n", raw_data.nunique())

# Display the dataset information
raw_data.info()

# Display the dataset descriptive statistics
raw_data.describe()

# Check for duplicates
raw_data.duplicated().sum()

### 2. Data Cleaning

- Handle missing values

- Rename columns

- Correct data types

- Remove duplicates



# keep only the columns needed for analysis
columns_to_keep = ["ProductID", "ProductName", "Description", "ProductBrand", "Gender", "Price (INR)", "PrimaryColor"]
df = raw_data[columns_to_keep].copy()

df


# Drop rows with missing values
df = df.dropna()

# Rename columns
df = df.rename(columns={
    "ProductID": "product_id",
    "ProductName": "product_name",
    "ProductBrand": "product_brand",
    "Description": "description",
    "Gender": "gender",
    "Price (INR)": "price_inr",
    "PrimaryColor": "primary_color"
})

# Ensure correct data types
df["product_id"]     = df["product_id"].astype("string")
df["product_name"]   = df["product_name"].astype("string")
df["product_brand"]  = df["product_brand"].astype("string")
df["description"]    = df["description"].astype("string")
df["gender"]         = df["gender"].astype("category")
df["primary_color"]  = df["primary_color"].astype("category")
df["price_inr"]      = df["price_inr"].astype("float64")

# Display the cleaned DataFrame
df.head()

# Remove spaces from string columns
string_columns = ["product_id", "product_name", "product_brand", "description", "gender", "primary_color"]

for col in string_columns:
    df[col] = df[col].str.strip()

# Ensure correct data types after cleaning
df["gender"]         = df["gender"].astype("category")
df["primary_color"]  = df["primary_color"].astype("category")

df.duplicated().sum()

# Display cleaned dataset information
df.info()

# Display cleaned dataset descriptive statistics
df.describe()

### 3. Save the cleaned data

df.to_csv("../data/product_cleaned.csv", index=False)
